// src/api/api.ts
export interface CommentVm {
    id?: string;
    content?: string | undefined;
    username?: string | undefined;
}
// ...
export interface PostInfo {
    title?: string | undefined;
    sellerLink?: string | undefined;
    description?: string | undefined;
    creationTime?: Date;
    lastEditTime?: Date;
}

export interface PostVm {
    id?: string;
    section?: SectionVm;
    info?: PostInfo;
    comments?: CommentVm[] | undefined;
    rate?: number;
    tags?: TagVm[] | undefined;
}

export interface ProblemDetails {
    type?: string | undefined;
    title?: string | undefined;
    status?: number | undefined;
    detail?: string | undefined;
    instance?: string | undefined;
}

export interface RatePostDto {
    postId?: string;
    rate?: number;
}

export interface SectionVm {
    id?: string;
    name?: string | undefined;
}

export interface TagVm {
    id?: string;
    name?: string | undefined;
}